Improved getMessages to parse data. Return said data as JSON to client requesting data with GET /messages. Renamed src/scripts to src/utilities for clarity
