+++ server/src/types/Message.ts
@@ -2,4 +2,5 @@ export interface Message {
     userId: string
     msgId: string
     msgContent: string
+    recievingUserIds: string
 }
